+++ internal/ui/settings.go
@@ -14,10 +14,10 @@ type deskSettings struct {
 	background             string
 	iconTheme              string
 	launcherIcons          []string
-	launcherIconSize       int
+	launcherIconSize       float32
 	launcherDisableTaskbar bool
 	launcherDisableZoom    bool
-	launcherZoomScale      float64
+	launcherZoomScale      float32
 	clockFormatting        string
 
 	moduleNames []string
@@ -38,7 +38,7 @@ func (d *deskSettings) LauncherIcons() []string {
 	return d.launcherIcons
 }
 
-func (d *deskSettings) LauncherIconSize() int {
+func (d *deskSettings) LauncherIconSize() float32 {
 	return d.launcherIconSize
 }
 
@@ -50,7 +50,7 @@ func (d *deskSettings) LauncherDisableZoom() bool {
 	return d.launcherDisableZoom
 }
 
-func (d *deskSettings) LauncherZoomScale() float64 {
+func (d *deskSettings) LauncherZoomScale() float32 {
 	return d.launcherZoomScale
 }
 
@@ -111,9 +111,9 @@ func (d *deskSettings) setLauncherIcons(defaultApps []string) {
 	d.apply()
 }
 
-func (d *deskSettings) setLauncherIconSize(size int) {
+func (d *deskSettings) setLauncherIconSize(size float32) {
 	d.launcherIconSize = size
-	fyne.CurrentApp().Preferences().SetInt("launchericonsize", d.launcherIconSize)
+	fyne.CurrentApp().Preferences().SetInt("launchericonsize", int(d.launcherIconSize))
 	d.apply()
 }
 
@@ -129,9 +129,9 @@ func (d *deskSettings) setLauncherDisableZoom(zoom bool) {
 	d.apply()
 }
 
-func (d *deskSettings) setLauncherZoomScale(scale float64) {
+func (d *deskSettings) setLauncherZoomScale(scale float32) {
 	d.launcherZoomScale = scale
-	fyne.CurrentApp().Preferences().SetFloat("launcherzoomscale", d.launcherZoomScale)
+	fyne.CurrentApp().Preferences().SetFloat("launcherzoomscale", float64(d.launcherZoomScale))
 	d.apply()
 }
 
@@ -177,7 +177,7 @@ func (d *deskSettings) load() {
 		}
 	}
 
-	d.launcherIconSize = fyne.CurrentApp().Preferences().Int("launchericonsize")
+	d.launcherIconSize = float32(fyne.CurrentApp().Preferences().Int("launchericonsize"))
 	if d.launcherIconSize == 0 {
 		d.launcherIconSize = 48
 	}
@@ -185,7 +185,7 @@ func (d *deskSettings) load() {
 	d.launcherDisableTaskbar = fyne.CurrentApp().Preferences().Bool("launcherdisabletaskbar")
 	d.launcherDisableZoom = fyne.CurrentApp().Preferences().Bool("launcherdisablezoom")
 
-	d.launcherZoomScale = fyne.CurrentApp().Preferences().Float("launcherzoomscale")
+	d.launcherZoomScale = float32(fyne.CurrentApp().Preferences().Float("launcherzoomscale"))
 	if d.launcherZoomScale == 0.0 {
 		d.launcherZoomScale = 2.0
 	}
