+++ internal/x11/wm/events.go
@@ -224,8 +224,8 @@ func (x *x11WM) handleMouseEnter(ev xproto.EnterNotifyEvent) {
 func (x *x11WM) handleMouseLeave(ev xproto.LeaveNotifyEvent) {
 	if mouseNotify, ok := fynedesk.Instance().(notify.MouseNotify); ok {
 		screen := fynedesk.Instance().Screens().ScreenForGeometry(int(ev.RootX), int(ev.RootY), 0, 0)
-		mouseNotify.MouseInNotify(fyne.NewPos(int(float32(ev.RootX)/screen.CanvasScale()),
-			int(float32(ev.RootY)/screen.CanvasScale())))
+		mouseNotify.MouseInNotify(fyne.NewPos(float32(ev.RootX)/screen.CanvasScale(),
+			float32(ev.RootY)/screen.CanvasScale()))
 	}
 }
