+++ settings.go
@@ -7,10 +7,10 @@ type DeskSettings interface {
 	ClockFormatting() string
 
 	LauncherIcons() []string
-	LauncherIconSize() int
+	LauncherIconSize() float32
 	LauncherDisableTaskbar() bool
 	LauncherDisableZoom() bool
-	LauncherZoomScale() float64
+	LauncherZoomScale() float32
 
 	ModuleNames() []string
