+++ wm/borderutil.go
@@ -3,6 +3,6 @@ package wm
 import "fyne.io/fynedesk"
 
 // ScaleToPixels calculates the pixels required to show a specified Fyne dimension on the given screen
-func ScaleToPixels(i int, screen *fynedesk.Screen) int {
-	return int(float32(i) * screen.CanvasScale())
+func ScaleToPixels(i float32, screen *fynedesk.Screen) int {
+	return int(i * screen.CanvasScale())
 }
